estimate cost fixed

+++ src/main/java/com/example/O_Way/dto/requestDto/RentalRequestDto.java
@@ -16,7 +16,7 @@ public class RentalRequestDto {
     private Long vehicleId;
 
     @NotNull
-    private Long distance;
+    private double distance;
 
     @NotNull
     private Double pickupLatitude;
